Unifying the planning component: Removing the helper class and changing the rules file according to UC suggestions.
