+++ development/pom.xml
@@ -26,7 +26,7 @@
                       <manifest>
                           <addClasspath>true</addClasspath>
                           <classpathPrefix>lib/</classpathPrefix>
-                          <mainClass>eubr.atmosphere.tma.Main</mainClass>
+                          <mainClass>eubr.atmosphere.tma.planning.Main</mainClass>
                       </manifest>
                   </archive>
                   <outputDirectory>${basedir}/bin</outputDirectory>
