--- development/src/main/resources/META-INF/kmodule.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-   Drools Online Course Sample Code and Study Materials (c) by Juhan Aasaru.
-
-   Drools Online Course Sample Code and Study Materials is licensed under a
-   Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 International License.
-
-   You should have received a copy of the license along with this
-   work. If not, see <http://creativecommons.org/licenses/by-nc-nd/4.0/>.
-  -->
-
-<kmodule
-  xmlns="http://www.drools.org/xsd/kmodule">
-
-  <kbase packages="eubr.atmosphere.tma.planning.privacy">
-    <ksession name="PrivacyRules"/>
-  </kbase>
-  <kbase packages="eubr.atmosphere.tma.planning.scalability">
-    <ksession name="Rules"/>
-  </kbase>
-
-</kmodule>
