+++ development/src/main/resources/eubr/atmosphere/tma/planning/Privacy.drl
@@ -1,0 +1,36 @@
+package eubr.atmosphere.tma.planning
+ 
+dialect "mvel"
+
+import eubr.atmosphere.tma.planning.AdaptationManager
+import eubr.atmosphere.tma.data.Action
+import eubr.atmosphere.tma.data.Configuration
+import eubr.atmosphere.tma.utils.TrustworthinessScore
+
+declare TrustworthinessScore
+      @role(event)
+      @expires(1s)
+      //@timestamp (eventDestinationTimestamp)
+ end
+
+rule "Privacy Score Validation - Increase PRIVAAS Anonymization"
+    when
+        $score: TrustworthinessScore ( privacyScore.score < privacyScore.getThreshold() )
+    then
+        Action action = new Action(4, "update", 8, 35001);
+	    action.addConfiguration(new Configuration(1, "score", $score.getPrivacyScore().getScore().toString()));
+	    action.addConfiguration(new Configuration(2, "atributeID", $score.getPrivacyScore().getAttributeId().toString()));
+	    action.addConfiguration(new Configuration(3, "configProfileID", $score.getPrivacyScore().getConfigurationProfileId().toString()));
+	    AdaptationManager.performAdaptation( action );
+end
+
+rule "Privacy Score Validation - Not Increase PRIVAAS Anonymization"
+    when
+        $score: TrustworthinessScore ( privacyScore.score >= privacyScore.getThreshold() )
+    then
+		Action action = new Action(4, "none", 8, 35001);
+	    action.addConfiguration(new Configuration(1, "score", $score.getPrivacyScore().getScore().toString()));
+	    action.addConfiguration(new Configuration(2, "atributeID", $score.getPrivacyScore().getAttributeId().toString()));
+	    action.addConfiguration(new Configuration(3, "configProfileID", $score.getPrivacyScore().getConfigurationProfileId().toString()));
+	    AdaptationManager.performAdaptation( action );
+end
