+++ development/src/main/resources/eubr/atmosphere/tma/planning/Rules.drl
