--- development/src/main/resources/eubr/atmosphere/tma/planning/privacy/PrivacyRules.drl
@@ -1,28 +1,0 @@
-package eubr.atmosphere.tma.planning
- 
-dialect "mvel"
-
-import eubr.atmosphere.tma.helpers.PrivacyHelper
-import eubr.atmosphere.tma.data.Action
-import eubr.atmosphere.tma.data.Configuration
-import eubr.atmosphere.tma.utils.PrivacyScore
-
-declare TrustworthinessScore
-      @role(event)
-      @expires(1s)
-      //@timestamp (eventDestinationTimestamp)
- end
-
-rule "Privacy Score Validation - Increase PRIVAAS Anonymization"
-    when
-        $score: PrivacyScore ( score < threshold )
-    then
-        PrivacyHelper.increasePRIVAASAnonymization( $score );
-end
-
-rule "Privacy Score Validation - Not Increase PRIVAAS Anonymization"
-    when
-        $score: PrivacyScore ( score >= threshold )
-    then
-        PrivacyHelper.noIncreasePRIVAASAnonymization( $score );
-end
